Add support for throttle groups within phases
Throttle groups allow you to execute specific jobs within a phase in
serial.  For example, if you have jobs A, B, and C, the default behavior
executes them in parallel, for example:

    A....
    B......
    C...

If we assign the same "throttle group" (a string) to jobs A and C, they
will execute in series (C will not start until A finishes).  This would
look like:

   A....
   B......
        C...

Throttle groups add another level of customization to MultiJobs.  They
support variable expansion and treat the "false" string as no throttle
group - so they can be easily attached to boolean build parameters.